Separate library-specific code into platforms/
This moves the emulator to a callback based approach for dealing with
platform specific code. The emulator needs two callbacks to run:

    - vblank_callback, which is called with the current framebuffer
      when the Gameboy enters vblank

    - should_close_callback, which is called each tick to determine
      if the Gameboy should keep running

This should allow this emulator to be easily ported to new graphics
libraries or systems.

This also enhances the build system to build the bulk of the code as a
static library, and to build one target per platform.