Fix sprite artifacts on the left hand side of the screen
An off-by-one error in sprite rendering was causing sprite data from one
pixel off the right side of the screen to be drawn (index 160, when the
maximum should be 159).

As the framebuffer is implemented as a single array, it happily
calculates this as being on the first column of the next row, causing
visual artifacts when sprites are positioned so they overlap the right
hand side of the screen.

To avoid bugs like this, we now use an 'is_on_screen' function which
encapsulates this logic and which is verified to be correct.