Fix run_test_roms
This was broken in 342c7569f936a15a5dc1f73a520f19372471412c when a flag
was introduced to stop printing serial output by default.

Since the test looks for 'failed' in the output and Blargg's tests write
that text to the serial output, the tests were passing trivially (by not
finding 'failed' in empty output). We pass the flag again to get output.

We also look for 'Passed' in the output and only pass if we find it,
which will prevent similar issues in the future.